+++ sql/lensed_agntruth/lensed_agn_schema.sql
@@ -9,5 +9,5 @@ CREATE TABLE lensed_agn_truth.truth_summary
 	coord earth);
 
 CREATE TABLE lensed_agn_truth.lensed_agn_variability_truth
-                  (id TEXT, obsHistID INTEGER, MJD REAL, bandpass TEXT,
-                  delta_flux REAL, num_photons REAL);
+                  (id TEXT, obsHistID INTEGER, MJD DOUBLE PRECISION,
+		  bandpass TEXT, delta_flux REAL, num_photons REAL);
